refactor: extract qr modal to own component
