+++ src/components/modals/index.ts
@@ -1,0 +1,3 @@
+import QrModal from "./QrModal.vue";
+
+export { QrModal };
