+++ src/components/wallet/Details.vue
@@ -4,7 +4,7 @@
     <div class="WalletHeaderDesktop hidden md:flex xl:rounded-lg">
       <button
         class="address-button ml-10 mr-6 p-3 rounded flex-none hover-button-shadow transition"
-        @click="toggleModal()"
+        @click="toggleModal"
       >
         <SvgIcon class="block" name="qr" view-box="0 0 29 29" />
       </button>
@@ -193,35 +193,7 @@
       </div>
     </div>
 
-    <!-- Modal -->
-    <Modal v-if="showModal" @close="toggleModal()">
-      <div :class="{ 'max-w-sm': !isCollapsed }" class="text-center px-10 py-2">
-        <p class="semibold text-3xl mb-4">
-          {{ $t("WALLET.QR_CODE") }}
-        </p>
-        <p class="mb-8">
-          {{ $t(`WALLET.SCAN_FOR_${isCollapsed ? "ADDRESS" : "URI"}`) }}
-        </p>
-
-        <div class="flex justify-between items-center">
-          <QrCode class="rounded" :value="qrValue" :options="{ size: 160 }" />
-          <div v-if="!isCollapsed">
-            <InputNumber :label="$t('TRANSACTION.AMOUNT')" name="amount" @input="onInputChange" />
-            <InputText :label="$t('TRANSACTION.SMARTBRIDGE')" name="vendorField" @input="onInputChange" />
-          </div>
-        </div>
-
-        <span
-          v-if="!isCollapsed"
-          class="block mt-4 whitespace-no-wrap overflow-x-auto rounded mx-auto bg-theme-feature-background px-4 py-2 text-white font-mono"
-          >{{ qrValue }}</span
-        >
-
-        <button class="mt-8 mx-auto pager-button items-center" @click="isCollapsed = !isCollapsed">
-          {{ $t(`WALLET.ADVANCED_QR.${isCollapsed ? "EXPAND" : "COLLAPSE"}`) }}
-        </button>
-      </div>
-    </Modal>
+    <QrModal v-if="showModal" :address="wallet.address" @close="toggleModal" />
   </section>
 </template>
 
@@ -230,14 +202,12 @@ import { Component, Prop, Vue, Watch } from "vue-property-decorator";
 import { mapGetters } from "vuex";
 import { IWallet } from "@/interfaces";
 import WalletVoters from "@/components/wallet/Voters.vue";
-import { InputText, InputNumber } from "@/components/search/input";
-import { URI_QRCODE_SCHEME_PREFIX } from "@/constants";
+import { QrModal } from "@/components/modals";
 
 @Component({
   components: {
-    InputNumber,
-    InputText,
     WalletVoters,
+    QrModal,
   },
   computed: {
     ...mapGetters("network", ["isListed", "knownWallets", "token"]),
@@ -253,9 +223,6 @@ export default class WalletDetails extends Vue {
   private isListed: boolean;
   private token: string;
   private currencyName: string;
-  private amount = 0;
-  private vendorField = "";
-  private isCollapsed = true;
 
   get name() {
     return this.knownWallets[this.wallet.address];
@@ -277,51 +244,16 @@ export default class WalletDetails extends Vue {
     return !!this.wallet.lockedBalance;
   }
 
-  get qrValue() {
-    return this.isCollapsed
-      ? this.wallet.address
-      : `${URI_QRCODE_SCHEME_PREFIX}${this.wallet.address}${this.formattedParams}`;
-  }
-
-  get formattedParams() {
-    const params = [];
-
-    if (this.amount) {
-      params.push(`amount=${this.amount}`);
-    }
-
-    if (this.vendorField) {
-      params.push(`vendorField=${this.vendorField}`);
-    }
-
-    return params.length > 0 ? `?${params.join("&")}` : "";
-  }
-
   get showBalanceTooltip() {
     return this.isListed && this.token !== this.currencyName;
-  }
-
-  @Watch("isCollapsed")
-  public onStatusChanged() {
-    this.amount = undefined;
-    this.vendorField = undefined;
-  }
-
-  private onInputChange(event: any) {
-    const { name, value } = event.target;
-    this[name] = value;
   }
 
   private setView(view: string) {
     this.view = view;
   }
 
-  private toggleModal() {
+  public toggleModal() {
     this.showModal = !this.showModal;
-
-    if (!this.showModal) {
-      this.isCollapsed = true;
-    }
   }
 }
 </script>
